+++ CJBCheatsMenu/Framework/Components/DescriptionElement.cs
@@ -1,3 +1,4 @@
+using System;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using StardewValley;
@@ -8,6 +9,13 @@ namespace CJBCheatsMenu.Framework.Components
     /// <summary>An options element which contains descriptive text.</summary>
     internal class DescriptionElement : BaseOptionsElement
     {
+        /*********
+        ** Fields
+        *********/
+        /// <summary>If set, get the label to display each tick.</summary>
+        private readonly Func<string>? GetLabel;
+
+
         /*********
         ** Accessors
         *********/
@@ -27,6 +35,15 @@ namespace CJBCheatsMenu.Framework.Components
             this.SplitLinesIfNeeded = splitLinesIfNeeded;
         }
 
+        /// <summary>Construct an instance.</summary>
+        /// <param name="label">The checkbox label.</param>
+        public DescriptionElement(Func<string> label)
+            : base(label(), -1, -1, 0, 0, 0)
+        {
+            this.GetLabel = label;
+            this.SplitLinesIfNeeded = false; // not compatible with dynamic line splitting, since that happens ahead of time
+        }
+
         /// <summary>Draw the component to the screen.</summary>
         /// <param name="spriteBatch">The sprite batch being drawn.</param>
         /// <param name="slotX">The X position at which to draw, relative to the bounds.</param>
@@ -34,6 +51,9 @@ namespace CJBCheatsMenu.Framework.Components
         /// <param name="context">The menu drawing the component.</param>
         public override void draw(SpriteBatch spriteBatch, int slotX, int slotY, IClickableMenu? context = null)
         {
+            if (this.GetLabel != null)
+                this.label = this.GetLabel();
+
             spriteBatch.DrawString(Game1.smallFont, this.label, new Vector2(slotX + this.bounds.X, slotY + this.bounds.Y), Color.Black);
         }
     }