+++ CJBCheatsMenu/docs/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## Upcoming release
 * Fixed 'freeze time' and 'grow' keybinds still triggered when typing into a text field.
+* Improved translations. Thanks to BernieAteUsername (updated Chinese), CaranudLapin (updated French), mitekano23 (updated Japanese), and MagoSupremo123 (updated Portuguese)!
 
 ## 1.36.2
 Released 15 April 2024 for SMAPI 4.0.0 or later.