+++ CJBCheatsMenu/docs/release-notes.md
@@ -8,6 +8,7 @@
 * You can now [make personal edits to the warp list](README.md#personal-warps).
 * Fixed 'freeze time' and 'grow' keybinds still triggered when typing into a text field.
 * Fixed missing translations for default tab names in Generic Mod Confgi Menu UI.
+* Tweaked weather tab's current-value format for clarity.
 * Improved translations. Thanks to BernieAteUsername (updated Chinese), CaranudLapin (updated French), mitekano23 (updated Japanese), and MagoSupremo123 (updated Portuguese)!
 
 ## 1.36.2