+++ CJBCheatsMenu/docs/release-notes.md
@@ -6,8 +6,10 @@
 * Added cheat to toggle perfection (thanks to PrimmR!).
 * Added keybind to reload settings from `config.json`.
 * You can now [make personal edits to the warp list](README.md#personal-warps).
+* You can now set tomorrow's weather to green rain.
 * Fixed 'freeze time' and 'grow' keybinds still triggered when typing into a text field.
-* Fixed missing translations for default tab names in Generic Mod Confgi Menu UI.
+* Fixed missing translations for default tab names in Generic Mod Config Menu UI.
+* Fixed unknown weathers shown as blank in the weather tab; it'll now show the weather ID instead.
 * Tweaked weather tab's current-value format for clarity.
 * Improved translations. Thanks to BernieAteUsername (updated Chinese), CaranudLapin (updated French), mitekano23 (updated Japanese), and MagoSupremo123 (updated Portuguese)!