+++ CJBCheatsMenu/docs/release-notes.md
@@ -10,6 +10,7 @@
 * Fixed 'freeze time' and 'grow' keybinds still triggered when typing into a text field.
 * Fixed missing translations for default tab names in Generic Mod Config Menu UI.
 * Fixed unknown weathers shown as blank in the weather tab; it'll now show the weather ID instead.
+* Fixed 'time frozen' message overlapping casino club coin display.
 * Tweaked weather tab's current-value format for clarity.
 * Improved translations. Thanks to BernieAteUsername (updated Chinese), CaranudLapin (updated French), mitekano23 (updated Japanese), and MagoSupremo123 (updated Portuguese)!